clean code

+++ tipi/com.dexels.navajo.tipi.jogl/src/com/dexels/navajo/dashboard/viewer/JOGLServerStatus.java
@@ -46,14 +46,23 @@ public class JOGLServerStatus {
 		this.postman = url;
 	}
 
+	/**
+	 * @param username  
+	 */
 	public void setUsername(String username) {
 
 	}
 
+	/**
+	 * @param password  
+	 */
 	public void setPassword(String password) {
 
 	}
 
+	/**
+	 * @param mode  
+	 */
 	public void draw(GL gl1, float width, float height, int mode) {
 		GL2 gl = gl1.getGL2();
 		gl.glBegin(GL2.GL_QUADS);
@@ -131,7 +140,7 @@ public class JOGLServerStatus {
 		
 		// Draw the index lines, ten, which is the number of max threads
 		int max_threads = 10;
-		float y_skip = (float)(max_height) / (float)max_threads;		
+		float y_skip = (max_height) / max_threads;		
 		gl.glBegin(GL.GL_LINES);
 		gl.glColor4f(0f, 0f, 0f, .4f);
 		for(int i=0;i<max_threads;i++){
@@ -242,6 +251,9 @@ public class JOGLServerStatus {
 
 	/*
 	 * Draw a string in OpenGL
+	 */
+	/**
+	 * @param gl  
 	 */
 	private void drawString(GL gl, String text, int xpos, int ypos, Color color, float alpha, float scale) {
 		float r = color.getRed() / 255f;
