+++ tipi/com.dexels.navajo.tipi.jogl/src/com/dexels/navajo/dashboard/viewer/JOGLTribeView.java
@@ -733,6 +733,9 @@ public class JOGLTribeView extends TipiDataComponentImpl implements GLEventListe
 		}
 	}
 
+	/**
+	 * @param gl  
+	 */
 	private final void compileStars(GL gl) {
 		try {
 			int size = 2000;
@@ -1219,7 +1222,7 @@ public class JOGLTribeView extends TipiDataComponentImpl implements GLEventListe
 		zoom_height = zoom / aspect_ratio;
 
 		/* create 5x5 pixel picking region near cursor location */
-		glu.gluPickMatrix((double) pickPoint.x, (double) (viewport[3] - pickPoint.y),// 
+		glu.gluPickMatrix(pickPoint.x, viewport[3] - pickPoint.y,// 
 				15.0, 15.0, viewport, 0);
 
 		// Zooming
@@ -1321,6 +1324,9 @@ public class JOGLTribeView extends TipiDataComponentImpl implements GLEventListe
 
 	/*
 	 * Draw a string in OpenGL
+	 */
+	/**
+	 * @param gl  
 	 */
 	private void drawString(GL gl, String text, int xpos, int ypos, Color color, float alpha, float scale) {
 		float r = color.getRed() / 255f;
